chat endpoints tested and fixed; chat refactor
- Added more verifications to AddFriendView
- Refactor a chat endpoint
- Moved every View class to each file inside a new views module
